+++ bonxai_core/include/bonxai/mask.hpp
@@ -139,9 +139,9 @@ class Mask {
   uint32_t WORD_COUNT;
 };
 
-//----------------------------------------------------------------
-//----------------------------------------------------------------
-//----------------------------------------------------------------
+//----------------------------------------------------
+//----------------- Implementations ------------------
+//----------------------------------------------------
 
 #define BONXAI_USE_INTRINSICS
